requirements.txt
streamlit
tensorflow-cpu
numpy
Pillow
rembg